+++ src/thread_pool_async.ml
@@ -2,83 +2,113 @@ open Base
 open Async_kernel
 open Async_unix
 
-type 'state thread = In_thread.Helper_thread.t * 'state
-
-type 'state t = {
-  pool : 'state thread Pipe.Reader.t * 'state thread Pipe.Writer.t;
-  create : unit -> 'state;
-  destroy : 'state -> unit;
-  threads : int;
+type 'state worker = {
+  thread : In_thread.Helper_thread.t;
+  state : 'state;
 }
 
-let init ~name ~threads ~create ~destroy =
-  let create_thread ~name ~create =
-    let%bind thread = In_thread.Helper_thread.create ~name () in
-    let%bind state = In_thread.run ~thread create in
-    return (thread, state)
-  in
-  (* No strict ordering on the heap. All items are equal *)
-  let reader, writer = Pipe.create () in
-  let elements = List.init threads ~f:ignore in
-  let add_to_pool () =
-    create_thread ~name ~create >>| Pipe.write_without_pushback writer
-  in
-  let%bind () = Deferred.List.iter elements ~how:`Parallel ~f:add_to_pool in
-  return {pool = reader, writer; create; destroy; threads}
+exception Invalid_thread_count of int
 
-type 'result computation =
-  [ `Ok of 'result
-  | `Attempt_retry ]
+exception Pool_already_destroyed
 
-let with'
-    :  'state t -> ?retries:int -> ('state -> 'result computation) ->
-    'result Deferred.Or_error.t
-  =
- fun {pool = reader, writer; create; destroy; _} ?(retries = 0) f ->
-  assert (retries >= 0);
-  let recreate ~thread state =
-    let%bind () = In_thread.run ~thread (fun () -> destroy state) in
-    In_thread.run ~thread create
+exception State_manipulation_error of exn
+
+type 'state t = {
+  reader : 'state worker Pipe.Reader.t;
+  writer : 'state worker Pipe.Writer.t;
+  create : unit -> 'state;
+  destroy : 'state -> unit;
+  thread_count : int;
+}
+
+let try_in_thread ~thread f = try_with (fun () -> In_thread.run ~thread f)
+
+let destroy_pool {reader; writer; thread_count; destroy; _} =
+  let%bind states_destroyed =
+    List.init thread_count ~f:ignore
+    |> Deferred.List.map ~how:`Parallel ~f:(fun () ->
+           match%bind Pipe.read reader with
+           | `Eof -> return @@ Ok ()
+           | `Ok {thread; state} -> try_in_thread ~thread (fun () -> destroy state))
+    |> Deferred.map ~f:Result.all_unit
   in
-  let run_once
-      : 'thread -> 'state -> ('result computation Or_error.t * 'state) Deferred.t
-    =
-   fun thread state ->
-    let%bind result =
-      In_thread.run ~thread (fun () -> Or_error.try_with (fun () -> f state))
+  Pipe.close writer;
+  match states_destroyed with
+  | Ok () -> return ()
+  | Error exn -> raise @@ State_manipulation_error exn
+
+let recreate_state ({create; destroy; _} as pool) {thread; state} =
+  let destroy_pool_and_raise exn =
+    let%bind () = destroy_pool pool in
+    raise @@ State_manipulation_error exn
+  in
+  match%bind try_in_thread ~thread (fun () -> destroy state) with
+  | Ok () -> (
+      match%bind try_in_thread ~thread create with
+      | Ok state -> return {thread; state}
+      | Error exn -> destroy_pool_and_raise exn)
+  | Error exn -> destroy_pool_and_raise exn
+
+let init ~name ~threads:thread_count ~create ~destroy =
+  match thread_count with
+  | n when n <= 0 -> raise @@ Invalid_thread_count thread_count
+  | _ -> (
+      (* No strict ordering on the heap. All items are equal *)
+      let reader, writer = Pipe.create () in
+      let pool = {reader; writer; create; destroy; thread_count} in
+      let%bind threads =
+        List.init thread_count ~f:ignore
+        |> Deferred.List.map ~how:`Parallel ~f:(In_thread.Helper_thread.create ~name)
+      in
+      let create_state thread = try_in_thread ~thread create in
+      match%bind
+        threads
+        |> Deferred.List.map ~how:`Parallel ~f:create_state
+        |> Deferred.map ~f:Result.all
+      with
+      | Ok states ->
+          let add_worker (thread, state) =
+            Pipe.write_without_pushback writer {thread; state}
+          in
+          List.zip_exn threads states |> List.iter ~f:add_worker;
+          return pool
+      | Error exn ->
+          let%bind () = destroy_pool pool in
+          raise @@ State_manipulation_error exn)
+
+let execute {thread; state} work = In_thread.run ~thread (fun () -> work state)
+
+let with_worker ({reader; writer; _} as pool) ?(retries = 0) job =
+  let run_once worker =
+    let%bind result = try_with (fun () -> job worker) in
+    let%bind worker =
+      match result with
+      | Ok (Ok _) -> return worker
+      | Ok (Error _) | Error _ -> recreate_state pool worker
     in
-    match result with
-    | Ok (`Ok res) -> return (Ok (`Ok res), state)
-    | Ok `Attempt_retry ->
-        let%bind state = recreate ~thread state in
-        return (Ok `Attempt_retry, state)
-    | Error _ as e ->
-        let%bind state = recreate ~thread state in
-        return (e, state)
+    return (result, worker)
+  in
+  let rec run_with_retry worker retries_left =
+    let%bind outcome, worker = run_once worker in
+    match outcome with
+    | Ok (Ok _) -> return (outcome, worker)
+    | Ok (Error _) -> (
+      match retries_left with
+      | n when n <= 0 -> return (outcome, worker)
+      | n -> run_with_retry worker (Int.pred n))
+    | Error _ -> return (outcome, worker)
   in
   match%bind Pipe.read reader with
-  | `Eof -> return @@ Or_error.errorf "Pool has been destroyed"
-  | `Ok (thread, state) ->
-      let rec run_with_retry state retries_left =
-        let%bind result, state = run_once thread state in
-        match result with
-        | Ok (`Ok result) -> return (Ok result, state)
-        | Ok `Attempt_retry -> (
-          match retries_left with
-          | n when n <= 0 -> return (Or_error.error_string "No more retries", state)
-          | n -> run_with_retry state (Int.pred n))
-        | Error err -> return (Error err, state)
-      in
-      let%bind result, state' = run_with_retry state retries in
-      Pipe.write_without_pushback writer (thread, state');
-      return result
+  | `Eof -> raise Pool_already_destroyed
+  | `Ok worker -> (
+      let%bind outcome, worker = run_with_retry worker retries in
+      Pipe.write_without_pushback writer worker;
+      match outcome with
+      | Ok outcome -> return outcome
+      | Error exn -> raise exn)
 
-let destroy {pool = reader, writer; threads; destroy; _} : unit Deferred.t =
-  let elements = List.init threads ~f:ignore in
-  let%bind () =
-    Deferred.List.iter elements ~how:`Parallel ~f:(fun () ->
-        match%bind Pipe.read reader with
-        | `Eof -> return ()
-        | `Ok (thread, state) -> In_thread.run ~thread (fun () -> destroy state))
-  in
-  return @@ Pipe.close writer
+let with' pool ?(retries = 0) work =
+  let job worker = execute worker work in
+  with_worker pool ~retries job
+
+let destroy = destroy_pool